+++ src/__tests__/applicationForms.int.test.ts
@@ -189,25 +189,25 @@ describe('/applicationForms', () => {
 			await createTestBaseFields();
 			await createApplicationFormField(db, null, {
 				applicationFormId: 3,
-				baseFieldId: 2,
+				baseFieldShortCode: 'yearsOfWork',
 				position: 1,
 				label: 'Anni Worki',
 			});
 			await createApplicationFormField(db, null, {
 				applicationFormId: 3,
-				baseFieldId: 1,
+				baseFieldShortCode: 'organizationName',
 				position: 2,
 				label: 'Org Nomen',
 			});
 			await createApplicationFormField(db, null, {
 				applicationFormId: 2,
-				baseFieldId: 1,
+				baseFieldShortCode: 'organizationName',
 				position: 2,
 				label: 'Name of Organization',
 			});
 			await createApplicationFormField(db, null, {
 				applicationFormId: 2,
-				baseFieldId: 2,
+				baseFieldShortCode: 'yearsOfWork',
 				position: 1,
 				label: 'Duration of work in years',
 			});
@@ -224,9 +224,8 @@ describe('/applicationForms', () => {
 					{
 						id: 4,
 						applicationFormId: 2,
-						baseFieldId: 2,
+						baseFieldShortCode: 'yearsOfWork',
 						baseField: {
-							id: 2,
 							label: 'Years of work',
 							description:
 								'The number of years the project will take to complete',
@@ -241,9 +240,8 @@ describe('/applicationForms', () => {
 					{
 						id: 3,
 						applicationFormId: 2,
-						baseFieldId: 1,
+						baseFieldShortCode: 'organizationName',
 						baseField: {
-							id: 1,
 							label: 'Organization Name',
 							description: 'The organizational name of the applicant',
 							shortCode: 'organizationName',
@@ -279,13 +277,13 @@ describe('/applicationForms', () => {
 			await createTestBaseFields();
 			await createApplicationFormField(db, null, {
 				applicationFormId: 1,
-				baseFieldId: 1,
+				baseFieldShortCode: 'organizationName',
 				position: 2,
 				label: 'Name of Organization',
 			});
 			await createApplicationFormField(db, null, {
 				applicationFormId: 1,
-				baseFieldId: 2,
+				baseFieldShortCode: 'yearsOfWork',
 				position: 1,
 				label: 'Duration of work in years',
 			});
@@ -300,11 +298,9 @@ describe('/applicationForms', () => {
 				version: 1,
 				fields: [
 					{
-						id: 2,
 						applicationFormId: 1,
-						baseFieldId: 2,
+						baseFieldShortCode: 'yearsOfWork',
 						baseField: {
-							id: 2,
 							label: 'Years of work',
 							description:
 								'The number of years the project will take to complete',
@@ -317,11 +313,9 @@ describe('/applicationForms', () => {
 						createdAt: expectTimestamp,
 					},
 					{
-						id: 1,
 						applicationFormId: 1,
-						baseFieldId: 1,
+						baseFieldShortCode: 'organizationName',
 						baseField: {
-							id: 1,
 							label: 'Organization Name',
 							description: 'The organizational name of the applicant',
 							shortCode: 'organizationName',
@@ -362,13 +356,13 @@ describe('/applicationForms', () => {
 			await createTestBaseFields();
 			await createApplicationFormField(db, null, {
 				applicationFormId: 1,
-				baseFieldId: 1,
+				baseFieldShortCode: 'organizationName',
 				position: 2,
 				label: 'Name of Organization',
 			});
 			await createApplicationFormField(db, null, {
 				applicationFormId: 1,
-				baseFieldId: 2,
+				baseFieldShortCode: 'yearsOfWork',
 				position: 1,
 				label: 'Duration of work in years',
 			});
@@ -510,9 +504,9 @@ describe('/applicationForms', () => {
 					opportunityId: '1',
 					fields: [
 						{
-							baseFieldId: '1',
+							baseFieldShortCode: 'organizationName',
 							position: 1,
-							label: 'Your First Name',
+							label: 'Organization Name',
 						},
 					],
 				})
@@ -527,10 +521,10 @@ describe('/applicationForms', () => {
 				fields: [
 					{
 						applicationFormId: 1,
-						baseFieldId: 1,
+						baseFieldShortCode: 'organizationName',
 						createdAt: expectTimestamp,
 						id: 1,
-						label: 'Your First Name',
+						label: 'Organization Name',
 						position: 1,
 					},
 				],
@@ -658,9 +652,9 @@ describe('/applicationForms', () => {
 					opportunityId: '1',
 					fields: [
 						{
-							baseFieldId: '1',
+							baseFieldShortCode: 'organizationName',
 							position: 1,
-							label: 'Your First Name',
+							label: 'Organization Name',
 						},
 					],
 				})