+++ src/__tests__/baseFields.int.test.ts
@@ -5,7 +5,6 @@ import {
 	createBaseField,
 	createOrUpdateBaseFieldLocalization,
 	loadBaseFieldLocalizationsBundleByBaseFieldId,
-	loadBaseFields,
 	loadTableMetrics,
 } from '../database';
 import { BaseFieldDataType, BaseFieldScope, PostgresErrorCode } from '../types';
@@ -32,12 +31,17 @@ const createTestBaseFieldWithLocalization = async () => {
 		dataType: BaseFieldDataType.STRING,
 		scope: BaseFieldScope.PROPOSAL,
 	});
-	await createOrUpdateBaseFieldLocalization(db, null, {
-		baseFieldId: baseField.id,
-		label: 'Le Resume',
-		description: 'Le Resume de la Applicant',
-		language: 'fr',
-	});
+	const baseFieldLocalization = await createOrUpdateBaseFieldLocalization(
+		db,
+		null,
+		{
+			baseFieldShortCode: baseField.shortCode,
+			label: 'Le Resume',
+			description: 'Le Resume de la Applicant',
+			language: 'fr',
+		},
+	);
+	return { baseField, baseFieldLocalization };
 };
 
 describe('/baseFields', () => {
@@ -67,14 +71,14 @@ describe('/baseFields', () => {
 			});
 
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: baseFieldOne.id,
+				baseFieldShortCode: baseFieldOne.shortCode,
 				language: 'fr',
 				label: 'prenom',
 				description: 'le prenom',
 			});
 
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: baseFieldTwo.id,
+				baseFieldShortCode: baseFieldTwo.shortCode,
 				language: 'fr',
 				label: 'postnom',
 				description: 'le postnom',
@@ -83,7 +87,6 @@ describe('/baseFields', () => {
 			const result = await request(app).get('/baseFields').expect(200);
 			expect(result.body).toMatchObject([
 				{
-					id: 1,
 					label: 'First Name',
 					description: 'The first name of the applicant',
 					shortCode: 'firstName',
@@ -94,14 +97,13 @@ describe('/baseFields', () => {
 							label: 'prenom',
 							language: 'fr',
 							createdAt: expectTimestamp,
-							baseFieldId: 1,
+							baseFieldShortCode: 'firstName',
 							description: 'le prenom',
 						},
 					},
 					createdAt: expectTimestamp,
 				},
 				{
-					id: 2,
 					label: 'Last Name',
 					description: 'The last name of the applicant',
 					shortCode: 'lastName',
@@ -112,7 +114,7 @@ describe('/baseFields', () => {
 							label: 'postnom',
 							language: 'fr',
 							createdAt: expectTimestamp,
-							baseFieldId: 2,
+							baseFieldShortCode: 'lastName',
 							description: 'le postnom',
 						},
 					},
@@ -148,7 +150,6 @@ describe('/baseFields', () => {
 			const after = await loadTableMetrics('base_fields');
 			expect(before.count).toEqual(0);
 			expect(result.body).toMatchObject({
-				id: expect.any(Number) as number,
 				label: '🏷️',
 				description: '😍',
 				shortCode: '🩳',
@@ -319,191 +320,14 @@ describe('/baseFields', () => {
 		});
 	});
 
-	describe('PUT /:baseFieldId', () => {
-		it('requires authentication', async () => {
-			await request(app).put('/baseFields/1').expect(401);
-		});
-
-		it('requires administrator role', async () => {
-			await request(app).put('/baseFields/1').set(authHeader).expect(401);
-		});
-
-		it('updates the specified base field', async () => {
-			// Not using the helper here because observing a change in values is explicitly
-			// the point of the test, so having full explicit control of the original value
-			// seems important.  Some day when we add better test tooling we can have it all.
-			await createBaseField(db, null, {
-				label: 'Summary',
-				description: 'A summary of the proposal',
-				shortCode: 'summary',
-				dataType: BaseFieldDataType.STRING,
-				scope: BaseFieldScope.PROPOSAL,
-			});
-			await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					shortCode: '🩳',
-					dataType: BaseFieldDataType.NUMBER,
-					scope: BaseFieldScope.ORGANIZATION,
-				})
-				.expect(200);
-			const baseFields = await loadBaseFields();
-			expect(baseFields[0]).toMatchObject({
-				id: 1,
-				label: '🏷️',
-				description: '😍',
-				shortCode: '🩳',
-				dataType: BaseFieldDataType.NUMBER,
-				scope: BaseFieldScope.ORGANIZATION,
-				localizations: {},
-				createdAt: expectTimestamp,
-			});
-		});
-
-		it('returns the updated base field', async () => {
-			await createTestBaseField();
-			const result = await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					shortCode: '🩳',
-					dataType: BaseFieldDataType.NUMBER,
-					scope: BaseFieldScope.ORGANIZATION,
-				})
-				.expect(200);
-			expect(result.body).toMatchObject({
-				id: 1,
-				label: '🏷️',
-				description: '😍',
-				shortCode: '🩳',
-				dataType: BaseFieldDataType.NUMBER,
-				scope: BaseFieldScope.ORGANIZATION,
-				localizations: {},
-				createdAt: expectTimestamp,
-			});
-		});
-
-		it('returns 400 bad request when no label is sent', async () => {
-			await createTestBaseField();
-
-			const result = await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					shortCode: '🩳',
-					description: '😍',
-					dataType: BaseFieldDataType.STRING,
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 400 bad request when no description is sent', async () => {
-			await createTestBaseField();
-			const result = await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					shortCode: '🩳',
-					dataType: BaseFieldDataType.STRING,
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 400 bad request when no shortCode is sent', async () => {
-			await createTestBaseField();
-			const result = await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					dataType: BaseFieldDataType.STRING,
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 400 bad request when no dataType is sent', async () => {
-			await createTestBaseField();
-			const result = await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					shortCode: '🩳',
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 400 when a non-numeric ID is sent', async () => {
-			const result = await request(app)
-				.put('/baseFields/notanumber')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					shortCode: 'firstName',
-					dataType: BaseFieldDataType.STRING,
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 404 when attempting to update a non-existent record', async () => {
-			await request(app)
-				.put('/baseFields/1')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: '🏷️',
-					description: '😍',
-					shortCode: '🩳',
-					dataType: BaseFieldDataType.STRING,
-					scope: BaseFieldScope.PROPOSAL,
-				})
-				.expect(404);
-		});
-	});
-	describe('GET /:baseFieldId/localizations', () => {
+	describe('GET /:baseFieldShortCode/localizations', () => {
 		it('does not require authentication', async () => {
 			await createTestBaseFieldWithLocalization();
-			await request(app).get('/baseFields/1/localizations').expect(200);
+			await request(app).get(`/baseFields/summary/localizations`).expect(200);
 		});
 
-		it('returns all base field localizations related to the given baseFieldId', async () => {
-			await createBaseField(db, null, {
+		it('returns all base field localizations related to the given baseFieldShortCode', async () => {
+			const baseField = await createBaseField(db, null, {
 				label: 'First Name',
 				description: 'The first name of the applicant',
 				shortCode: 'firstName',
@@ -512,34 +336,34 @@ describe('/baseFields', () => {
 			});
 
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				language: 'fr',
 				label: 'prenom',
 				description: 'le prenom',
 			});
 
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				language: 'en',
 				label: 'First Name',
 				description: 'The First Name of the applicant',
 			});
 
 			const result = await request(app)
-				.get('/baseFields/1/localizations')
+				.get(`/baseFields/${baseField.shortCode}/localizations`)
 				.expect(200);
 			expect(result.body).toMatchObject({
 				total: 2,
 				entries: [
 					{
-						baseFieldId: 1,
+						baseFieldShortCode: baseField.shortCode,
 						language: 'fr',
 						label: 'prenom',
 						description: 'le prenom',
 						createdAt: expectTimestamp,
 					},
 					{
-						baseFieldId: 1,
+						baseFieldShortCode: baseField.shortCode,
 						language: 'en',
 						label: 'First Name',
 						description: 'The First Name of the applicant',
@@ -550,7 +374,7 @@ describe('/baseFields', () => {
 		});
 		it('returns 404 when a base field is referenced that does not exist', async () => {
 			const result = await request(app)
-				.get('/baseFields/1/localizations')
+				.get('/baseFields/fakeShortCode/localizations')
 				.expect(404);
 			expect(result.body).toMatchObject({
 				name: 'NotFoundError',
@@ -579,7 +403,7 @@ describe('/baseFields', () => {
 			const testBaseField = await createTestBaseField();
 			const before = await loadTableMetrics('base_field_localizations');
 			await request(app)
-				.put('/baseFields/1/localizations/fr')
+				.put(`/baseFields/${testBaseField.shortCode}/localizations/fr`)
 				.type('application/json')
 				.set(adminUserAuthHeader)
 				.send({
@@ -592,13 +416,13 @@ describe('/baseFields', () => {
 				await loadBaseFieldLocalizationsBundleByBaseFieldId(
 					db,
 					null,
-					testBaseField.id,
+					testBaseField.shortCode,
 					NO_LIMIT,
 					NO_OFFSET,
 				);
 			expect(before.count).toEqual(0);
 			expect(baseFieldLocalizations.entries[0]).toMatchObject({
-				baseFieldId: 1,
+				baseFieldShortCode: testBaseField.shortCode,
 				label: 'Résume',
 				description: 'Le Résume de proposal',
 				createdAt: expectTimestamp,
@@ -607,22 +431,22 @@ describe('/baseFields', () => {
 		});
 
 		it('updates only the specified base field if it does exist', async () => {
-			const testBaseField = await createTestBaseField();
+			const baseField = await createTestBaseField();
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				language: 'fr',
 				label: 'Résume',
 				description: 'Le Résume de proposal',
 			});
 			await createOrUpdateBaseFieldLocalization(db, null, {
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				language: 'en',
 				label: 'Summary',
 				description: 'The Summary of a proposal',
 			});
 			const before = await loadTableMetrics('base_field_localizations');
 			await request(app)
-				.put('/baseFields/1/localizations/fr')
+				.put(`/baseFields/${baseField.shortCode}/localizations/fr`)
 				.type('application/json')
 				.set(adminUserAuthHeader)
 				.send({
@@ -635,19 +459,19 @@ describe('/baseFields', () => {
 				await loadBaseFieldLocalizationsBundleByBaseFieldId(
 					db,
 					null,
-					testBaseField.id,
+					baseField.shortCode,
 					NO_LIMIT,
 					NO_OFFSET,
 				);
 			expect(before.count).toEqual(2);
 			expect(baseFieldLocalizations.entries[0]).toMatchObject({
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				label: 'Le Résume',
 				description: 'Le grand Résume de proposal',
 				createdAt: expectTimestamp,
 			});
 			expect(baseFieldLocalizations.entries[1]).toMatchObject({
-				baseFieldId: 1,
+				baseFieldShortCode: baseField.shortCode,
 				label: 'Summary',
 				description: 'The Summary of a proposal',
 				createdAt: expectTimestamp,
@@ -659,7 +483,7 @@ describe('/baseFields', () => {
 			await createTestBaseField();
 
 			const result = await request(app)
-				.put('/baseFields/1/localizations/fr')
+				.put('/baseFields/summary/localizations/fr')
 				.type('application/json')
 				.set(adminUserAuthHeader)
 				.send({
@@ -675,27 +499,11 @@ describe('/baseFields', () => {
 		it('returns 400 bad request when no description is sent', async () => {
 			await createTestBaseField();
 			const result = await request(app)
-				.put('/baseFields/1/localizations/fr')
+				.put('/baseFields/summary/localizations/fr')
 				.type('application/json')
 				.set(adminUserAuthHeader)
 				.send({
 					label: 'Résume',
-				})
-				.expect(400);
-			expect(result.body).toMatchObject({
-				name: 'InputValidationError',
-				details: expect.any(Array) as unknown[],
-			});
-		});
-
-		it('returns 400 when a non-numeric ID is sent', async () => {
-			const result = await request(app)
-				.put('/baseFields/notanumber/localizations/fr')
-				.type('application/json')
-				.set(adminUserAuthHeader)
-				.send({
-					label: 'Résume',
-					description: 'Le Résume de proposal',
 				})
 				.expect(400);
 			expect(result.body).toMatchObject({
@@ -708,7 +516,7 @@ describe('/baseFields', () => {
 			await createTestBaseField();
 			const result = await request(app)
 				.put(
-					'/baseFields/1/localizations/theLanguageKlingonWhichIsNotARealLanguage',
+					'/baseFields/summary/localizations/theLanguageKlingonWhichIsNotARealLanguage',
 				)
 				.type('application/json')
 				.set(adminUserAuthHeader)
@@ -725,7 +533,7 @@ describe('/baseFields', () => {
 
 		it('returns 404 when a base field is referenced that does not exist', async () => {
 			const result = await request(app)
-				.put('/baseFields/1/localizations/fr')
+				.put('/baseFields/shortCode/localizations/fr')
 				.type('application/json')
 				.set(adminUserAuthHeader)
 				.send({