+++ src/database/queries/baseFields/updateById.sql
@@ -1,8 +1,0 @@
-UPDATE base_fields SET
-	label = :label,
-	description = :description,
-	short_code = :shortCode,
-	data_type = :dataType,
-	scope = :scope
-WHERE id = :baseFieldId
-RETURNING base_field_to_json(base_fields) AS object;