+++ src/routers/baseFieldsRouter.ts
@@ -10,17 +10,12 @@ baseFieldsRouter.post(
 	requireAdministratorRole,
 	baseFieldsHandlers.postBaseField,
 );
-baseFieldsRouter.put(
-	'/:baseFieldId',
-	requireAdministratorRole,
-	baseFieldsHandlers.putBaseField,
-);
 baseFieldsRouter.get(
-	'/:baseFieldId/localizations',
-	baseFieldsHandlers.getBaseFieldLocalizationsByBaseFieldId,
+	'/:baseFieldShortCode/localizations',
+	baseFieldsHandlers.getBaseFieldLocalizationsByBaseFieldShortCode,
 );
 baseFieldsRouter.put(
-	'/:baseFieldId/localizations/:language',
+	'/:baseFieldShortCode/localizations/:language',
 	requireAdministratorRole,
 	baseFieldsHandlers.putBaseFieldLocalization,
 );