+++ src/tasks/__tests__/copyBaseFields.int.test.ts
@@ -39,20 +39,19 @@ const createTestBaseFieldsCopyTask = async (
 };
 
 const mockFirstNameBaseField = {
-	id: 1,
 	label: 'First Name',
 	description: 'The first name of the applicant',
 	shortCode: 'first_name',
 	localizations: {
 		en: {
-			baseFieldId: 1,
+			baseFieldShortCode: 'first_name',
 			language: 'en',
 			label: 'First name',
 			description: 'The first name of the Applicant',
 			createdAt: '2024-10-17T01:46:58.494Z',
 		},
 		fr: {
-			baseFieldId: 1,
+			baseFieldShortCode: 'first_name',
 			language: 'fr',
 			label: 'Le Prenom',
 			description: 'Le Prenom de la Applicant',
@@ -65,7 +64,6 @@ const mockFirstNameBaseField = {
 };
 
 const mockFirstNameBaseFieldWithNoLocalizations = {
-	id: 1,
 	label: 'First Name',
 	description: 'The first name of the applicant',
 	shortCode: 'first_name',
@@ -76,20 +74,19 @@ const mockFirstNameBaseFieldWithNoLocalizations = {
 };
 
 const mockLastNameBaseField = {
-	id: 2,
 	label: 'Last Name',
 	description: 'The last name of the applicant',
 	shortCode: 'last_name',
 	localizations: {
 		en: {
-			baseFieldId: 2,
+			baseFieldShortCode: 'last_name',
 			language: 'en',
 			label: 'Last name',
 			description: 'The Last name of the Applicant',
 			createdAt: '2024-10-17T01:46:58.494Z',
 		},
 		fr: {
-			baseFieldId: 2,
+			baseFieldShortCode: 'last_name',
 			language: 'fr',
 			label: 'Nom de famille',
 			description: 'Le nom de la famille de la applicant',
@@ -322,23 +319,26 @@ describe('copyBaseFields', () => {
 		expect(before.count).toEqual(0);
 		expect(after.count).toEqual(1);
 
-		const insertedRemoteBaseField = await loadBaseField(db, null, 1);
+		const insertedRemoteBaseField = await loadBaseField(
+			db,
+			null,
+			mockFirstNameBaseField.shortCode,
+		);
 
 		expect(insertedRemoteBaseField).toEqual({
-			id: 1,
 			label: 'First Name',
 			description: 'The first name of the applicant',
 			shortCode: mockFirstNameBaseField.shortCode,
 			localizations: {
 				en: {
-					baseFieldId: 1,
+					baseFieldShortCode: mockFirstNameBaseField.shortCode,
 					language: 'en',
 					label: 'First name',
 					description: 'The first name of the Applicant',
 					createdAt: expectTimestamp,
 				},
 				fr: {
-					baseFieldId: 1,
+					baseFieldShortCode: mockFirstNameBaseField.shortCode,
 					language: 'fr',
 					label: 'Le Prenom',
 					description: 'Le Prenom de la Applicant',
@@ -395,7 +395,7 @@ describe('copyBaseFields', () => {
 		const localBaseFieldAfterInsertion = await loadBaseField(
 			db,
 			null,
-			localBaseField.id,
+			localBaseField.shortCode,
 		);
 
 		expect(localBaseFieldAfterInsertion).toEqual(localBaseField);
@@ -451,16 +451,19 @@ describe('copyBaseFields', () => {
 
 		const after = await loadTableMetrics('base_fields');
 
-		const updatedBaseField = await loadBaseField(db, null, localBaseField.id);
+		const updatedBaseField = await loadBaseField(
+			db,
+			null,
+			localBaseField.shortCode,
+		);
 
 		expect(before.count).toEqual(1);
 		expect(after.count).toEqual(1);
 
 		expect(updatedBaseField).toEqual({
-			id: localBaseField.id,
 			label: 'Local Data',
 			description: 'This is local data',
-			shortCode: 'ld',
+			shortCode: localBaseField.shortCode,
 			dataType: 'string',
 			scope: 'proposal',
 			createdAt: localBaseField.createdAt,
@@ -518,39 +521,45 @@ describe('copyBaseFields', () => {
 
 		const after = await loadTableMetrics('base_fields');
 
-		const updatedBaseField = await loadBaseField(db, null, localBaseField.id);
+		const updatedBaseField = await loadBaseField(
+			db,
+			null,
+			localBaseField.shortCode,
+		);
 
 		expect(before.count).toEqual(1);
 		expect(after.count).toEqual(2);
 
 		expect(updatedBaseField).toEqual({
-			id: localBaseField.id,
 			label: 'Local Data',
 			description: 'This is local data',
-			shortCode: 'ld',
+			shortCode: localBaseField.shortCode,
 			dataType: 'string',
 			scope: 'proposal',
 			createdAt: localBaseField.createdAt,
 			localizations: {},
 		});
 
-		const insertedRemoteBaseField = await loadBaseField(db, null, 3);
+		const insertedRemoteBaseField = await loadBaseField(
+			db,
+			null,
+			mockFirstNameBaseField.shortCode,
+		);
 
 		expect(insertedRemoteBaseField).toEqual({
-			id: 3,
 			label: 'First Name',
 			description: 'The first name of the applicant',
 			shortCode: mockFirstNameBaseField.shortCode,
 			localizations: {
 				en: {
-					baseFieldId: 3,
+					baseFieldShortCode: mockFirstNameBaseField.shortCode,
 					language: 'en',
 					label: 'First name',
 					description: 'The first name of the Applicant',
 					createdAt: expectTimestamp,
 				},
 				fr: {
-					baseFieldId: 3,
+					baseFieldShortCode: mockFirstNameBaseField.shortCode,
 					language: 'fr',
 					label: 'Le Prenom',
 					description: 'Le Prenom de la Applicant',
@@ -576,7 +585,7 @@ describe('copyBaseFields', () => {
 		});
 
 		await createOrUpdateBaseFieldLocalization(db, null, {
-			baseFieldId: localBaseField.id,
+			baseFieldShortCode: localBaseField.shortCode,
 			label: 'Le Prenom',
 			description: 'Le Prenom de la Applicant',
 			language: 'fr',
@@ -606,16 +615,19 @@ describe('copyBaseFields', () => {
 			baseFieldsCopyTask.id,
 		);
 
-		const updatedBaseField = await loadBaseField(db, null, localBaseField.id);
+		const updatedBaseField = await loadBaseField(
+			db,
+			null,
+			localBaseField.shortCode,
+		);
 
 		expect(updatedBaseField).toEqual({
-			id: localBaseField.id,
 			label: 'First Name',
 			description: 'The first name of the applicant',
 			shortCode: mockFirstNameBaseField.shortCode,
 			localizations: {
 				fr: {
-					baseFieldId: localBaseField.id,
+					baseFieldShortCode: localBaseField.shortCode,
 					language: 'fr',
 					label: 'Le Prenom',
 					description: 'Le Prenom de la Applicant',
@@ -641,7 +653,7 @@ describe('copyBaseFields', () => {
 		});
 
 		await createOrUpdateBaseFieldLocalization(db, null, {
-			baseFieldId: localBaseField.id,
+			baseFieldShortCode: localBaseField.shortCode,
 			label: 'Nombre de Pila',
 			description: 'Nombre de Pila',
 			language: 'sp',
@@ -671,30 +683,33 @@ describe('copyBaseFields', () => {
 			baseFieldsCopyTask.id,
 		);
 
-		const updatedBaseField = await loadBaseField(db, null, localBaseField.id);
+		const updatedBaseField = await loadBaseField(
+			db,
+			null,
+			localBaseField.shortCode,
+		);
 
 		expect(updatedBaseField).toEqual({
-			id: localBaseField.id,
 			label: 'First Name',
 			description: 'The first name of the applicant',
 			shortCode: mockFirstNameBaseField.shortCode,
 			localizations: {
 				en: {
-					baseFieldId: localBaseField.id,
+					baseFieldShortCode: localBaseField.shortCode,
 					language: 'en',
 					label: 'First name',
 					description: 'The first name of the Applicant',
 					createdAt: expectTimestamp,
 				},
 				fr: {
-					baseFieldId: localBaseField.id,
+					baseFieldShortCode: localBaseField.shortCode,
 					language: 'fr',
 					label: 'Le Prenom',
 					description: 'Le Prenom de la Applicant',
 					createdAt: expectTimestamp,
 				},
 				sp: {
-					baseFieldId: localBaseField.id,
+					baseFieldShortCode: localBaseField.shortCode,
 					label: 'Nombre de Pila',
 					description: 'Nombre de Pila',
 					language: 'sp',
@@ -780,26 +795,25 @@ describe('copyBaseFields', () => {
 
 		const after = await loadTableMetrics('base_fields');
 
-		const updatedBaseField = await loadBaseField(db, null, baseField.id);
+		const updatedBaseField = await loadBaseField(db, null, baseField.shortCode);
 
 		expect(before.count).toEqual(1);
 		expect(after.count).toEqual(1);
 
 		expect(updatedBaseField).toEqual({
-			id: baseField.id,
 			label: 'First Name',
 			description: 'The first name of the applicant',
-			shortCode: 'first_name',
+			shortCode: baseField.shortCode,
 			localizations: {
 				en: {
-					baseFieldId: baseField.id,
+					baseFieldShortCode: baseField.shortCode,
 					language: 'en',
 					label: 'First name',
 					description: 'The first name of the Applicant',
 					createdAt: expectTimestamp,
 				},
 				fr: {
-					baseFieldId: 1,
+					baseFieldShortCode: baseField.shortCode,
 					language: 'fr',
 					label: 'Le Prenom',
 					description: 'Le Prenom de la Applicant',