+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.18.0 2025-02-13
+
+### Changed
+
+- `BaseFields` are now primary keyed on the `ShortCode` value, and have had corresponding
+  foreign keys in the `ApplicationFormField` and `BaseFieldLocalization` tables updated to match.
+
 ## 0.17.0 2025-02-13
 
 ### Changed